added ApiResponse Format + Swagger UI

+++ src/main/java/com/springRest/bookommerce/services/BookServiceImpl.java
@@ -1,12 +1,16 @@
 package com.springRest.bookommerce.services;
 
+import com.springRest.bookommerce.dto.ApiResponse;
 import com.springRest.bookommerce.model.BookModel;
 import com.springRest.bookommerce.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class BookServiceImpl implements BookService{
@@ -15,24 +19,38 @@ public class BookServiceImpl implements BookService{
     private BookRepository bookRepository;
 
     @Override
-    public List<BookModel> getBookList() {
-        return bookRepository.findAll();
+    public ResponseEntity<ApiResponse<List<BookModel>>> getBookList() {
+        List<BookModel> bookModelList = bookRepository.findAll();
+        if (bookModelList.isEmpty()){
+            return new ResponseEntity<>(new ApiResponse<>(200,"No Book Found",bookModelList), HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(new ApiResponse<>(200,"Book Found",bookModelList), HttpStatus.OK);
+        }
     }
 
     @Override
-    public BookModel addBook(BookModel bookModel) {
-        bookRepository.save(bookModel);
-        return bookModel;
+    public ResponseEntity<ApiResponse<BookModel>> addBook(BookModel bookModel) {
+        BookModel book = bookRepository.save(bookModel);
+        return new ResponseEntity<>(new ApiResponse<>(201,"Book Added",book),HttpStatus.CREATED);
     }
 
     @Override
-    public String deleteBook(Long id) {
-        bookRepository.deleteById(id);
-        return "Book Deleted Successful!";
+    public ResponseEntity<ApiResponse<String>> deleteBook(Long id) {
+        Optional<BookModel> bookModelOptional = bookRepository.findById(id);
+        if (bookModelOptional.isPresent()){
+            bookRepository.deleteById(id);
+            return new ResponseEntity<>(new ApiResponse<>(200,"Book Deleted Successfully!",null),HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(new ApiResponse<>(200,"Book Id Not Found!",null),HttpStatus.OK);
+        }
+
+
     }
 
     @Override
-    public BookModel editBook(Long bookId, BookModel bookModel) {
+    public ResponseEntity<ApiResponse<BookModel>> editBook(Long bookId, BookModel bookModel) {
         BookModel objectBookModel = bookRepository.findById(bookId).get();
 
         if (Objects.nonNull(bookModel.getBookName()) && !"".equalsIgnoreCase(bookModel.getBookName())){
@@ -45,6 +63,14 @@ public class BookServiceImpl implements BookService{
             objectBookModel.setBookPrice(bookModel.getBookPrice());
         }
 
-        return bookRepository.save(objectBookModel);
+        Optional<BookModel> bookModelOptional = bookRepository.findById(bookId);
+        if (bookModelOptional.isPresent()){
+            bookRepository.save(objectBookModel);
+            return new ResponseEntity<>(new ApiResponse<>(200,"Book Edit Successful!",objectBookModel),HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(new ApiResponse<>(200,"Book Id Not Found!",null),HttpStatus.OK);
+        }
+
     }
 }
